kdfkjdf

+++ community/primitive-collections/src/main/java/org/neo4j/collection/primitive/PrimitiveIntCollections.java
@@ -607,9 +607,9 @@ public class PrimitiveIntCollections
         return set;
     }
 
-    public static int count( PrimitiveIntIterator iterator )
+    public static long count( PrimitiveIntIterator iterator )
     {
-        int count = 0;
+        long count = 0;
         for ( ; iterator.hasNext(); iterator.next(), count++ )
         {   // Just loop through this
         }
